added build_unauth method to ActionPlan enum
